+++ modules/globebrowsing/src/globelabelscomponent.cpp
@@ -357,7 +357,9 @@ void GlobeLabelsComponent::initialize(const ghoul::Dictionary& dictionary,
     }
 
     if (dictionary.hasKey(LabelsFontSizeInfo.identifier)) {
-        _labelsFontSize = dictionary.value<float>(LabelsFontSizeInfo.identifier);
+        _labelsFontSize = static_cast<float>(
+            dictionary.value<double>(LabelsFontSizeInfo.identifier)
+        );
         _labelsFontSize.onChange([this]() { initializeFonts(); });
     }
 
@@ -368,15 +370,19 @@ void GlobeLabelsComponent::initialize(const ghoul::Dictionary& dictionary,
     }
 
     if (dictionary.hasKey(LabelsMinHeightInfo.identifier)) {
-        _labelsMinHeight = dictionary.value<float>(LabelsMinHeightInfo.identifier);
+        _labelsMinHeight = static_cast<float>(
+            dictionary.value<double>(LabelsMinHeightInfo.identifier)
+        );
     }
 
     if (dictionary.hasKey(LabelsColorInfo.identifier)) {
-        _labelsColor = dictionary.value<glm::vec3>(LabelsColorInfo.identifier);
+        _labelsColor = dictionary.value<glm::dvec3>(LabelsColorInfo.identifier);
     }
 
     if (dictionary.hasKey(LabelsOpacityInfo.identifier)) {
-        _labelsOpacity = dictionary.value<float>(LabelsOpacityInfo.identifier);
+        _labelsOpacity = static_cast<float>(
+            dictionary.value<double>(LabelsOpacityInfo.identifier)
+        );
     }
 
     if (dictionary.hasKey(LabelsFadeInEnabledInfo.identifier)) {
@@ -384,8 +390,8 @@ void GlobeLabelsComponent::initialize(const ghoul::Dictionary& dictionary,
     }
 
     if (dictionary.hasKey(LabelsFadeInStartingDistanceInfo.identifier)) {
-        _labelsFadeInDist = dictionary.value<float>(
-            LabelsFadeInStartingDistanceInfo.identifier
+        _labelsFadeInDist = static_cast<float>(
+            dictionary.value<double>(LabelsFadeInStartingDistanceInfo.identifier)
         );
     }
 
@@ -396,20 +402,20 @@ void GlobeLabelsComponent::initialize(const ghoul::Dictionary& dictionary,
     }
 
     if (dictionary.hasKey(LabelsFadeOutStartingDistanceInfo.identifier)) {
-        _labelsFadeOutDist = dictionary.value<float>(
-            LabelsFadeOutStartingDistanceInfo.identifier
+        _labelsFadeOutDist = static_cast<float>(
+            dictionary.value<double>(LabelsFadeOutStartingDistanceInfo.identifier)
         );
     }
 
     if (dictionary.hasKey(LabelsMinSizeInfo.identifier)) {
         _labelsMinSize = static_cast<int>(
-            dictionary.value<float>(LabelsMinSizeInfo.identifier)
+            dictionary.value<double>(LabelsMinSizeInfo.identifier)
         );
     }
 
     if (dictionary.hasKey(LabelsMaxSizeInfo.identifier)) {
         _labelsMaxSize = static_cast<int>(
-            dictionary.value<float>(LabelsMaxSizeInfo.identifier)
+            dictionary.value<double>(LabelsMaxSizeInfo.identifier)
         );
     }
