+++ modules/globebrowsing/src/layeradjustment.cpp
@@ -134,13 +134,16 @@ void LayerAdjustment::setValuesFromDictionary(const ghoul::Dictionary& adjustmen
         );
     }
 
-    if (adjustmentDict.hasKeyAndValue<glm::vec3>(KeyChromaKeyColor)) {
-        glm::vec3 dictChromaKeyColor = adjustmentDict.value<glm::vec3>(KeyChromaKeyColor);
+    if (adjustmentDict.hasKeyAndValue<glm::dvec3>(KeyChromaKeyColor)) {
+        glm::vec3 dictChromaKeyColor =
+            adjustmentDict.value<glm::dvec3>(KeyChromaKeyColor);
         _chromaKeyColor = std::move(dictChromaKeyColor);
     }
 
-    if (adjustmentDict.hasKeyAndValue<float>(KeyChromaKeyTolerance)) {
-        float dictChromaKeyTolerance = adjustmentDict.value<float>(KeyChromaKeyTolerance);
+    if (adjustmentDict.hasKeyAndValue<double>(KeyChromaKeyTolerance)) {
+        float dictChromaKeyTolerance = static_cast<float>(
+            adjustmentDict.value<double>(KeyChromaKeyTolerance)
+        );
         _chromaKeyTolerance = dictChromaKeyTolerance;
     }
 }
