+++ modules/kameleonvolume/rendering/renderablekameleonvolume.cpp
@@ -154,15 +154,15 @@ RenderableKameleonVolume::RenderableKameleonVolume(const ghoul::Dictionary& dict
     , _transferFunctionPath(TransferFunctionInfo)
     , _cache(CacheInfo)
 {
-    if (dictionary.hasKeyAndValue<glm::vec3>(KeyDimensions)) {
-        _dimensions = dictionary.value<glm::vec3>(KeyDimensions);
+    if (dictionary.hasKeyAndValue<glm::dvec3>(KeyDimensions)) {
+        _dimensions = dictionary.value<glm::dvec3>(KeyDimensions);
     }
     else {
         LWARNING("No dimensions specified for volumetric data, falling back to 32^3");
         _dimensions = glm::uvec3(32);
     }
 
-    _stepSize = dictionary.value<float>(KeyStepSize);
+    _stepSize = static_cast<float>(dictionary.value<double>(KeyStepSize));
 
     if (dictionary.hasKeyAndValue<std::string>(KeyTransferFunction)) {
         _transferFunctionPath = dictionary.value<std::string>(KeyTransferFunction);
@@ -179,33 +179,37 @@ RenderableKameleonVolume::RenderableKameleonVolume(const ghoul::Dictionary& dict
         _variable = dictionary.value<std::string>(KeyVariable);
     }
 
-    if (dictionary.hasKeyAndValue<glm::vec3>(KeyLowerDomainBound)) {
-        _lowerDomainBound = dictionary.value<glm::vec3>(KeyLowerDomainBound);
+    if (dictionary.hasKeyAndValue<glm::dvec3>(KeyLowerDomainBound)) {
+        _lowerDomainBound = dictionary.value<glm::dvec3>(KeyLowerDomainBound);
     }
     else {
         _autoDomainBounds = true;
     }
 
-    if (dictionary.hasKeyAndValue<glm::vec3>(KeyUpperDomainBound)) {
-        _upperDomainBound = dictionary.value<glm::vec3>(KeyUpperDomainBound);
+    if (dictionary.hasKeyAndValue<glm::dvec3>(KeyUpperDomainBound)) {
+        _upperDomainBound = dictionary.value<glm::dvec3>(KeyUpperDomainBound);
     }
     else {
         _autoDomainBounds = true;
     }
 
-    if (dictionary.hasKeyAndValue<glm::vec3>(KeyDomainScale)) {
-        _domainScale = dictionary.value<glm::vec3>(KeyDomainScale);
+    if (dictionary.hasKeyAndValue<glm::dvec3>(KeyDomainScale)) {
+        _domainScale = dictionary.value<glm::dvec3>(KeyDomainScale);
     }
 
-    if (dictionary.hasKeyAndValue<float>(KeyLowerValueBound)) {
-        _lowerValueBound = dictionary.value<float>(KeyLowerValueBound);
+    if (dictionary.hasKeyAndValue<double>(KeyLowerValueBound)) {
+        _lowerValueBound = static_cast<float>(
+            dictionary.value<double>(KeyLowerValueBound)
+        );
     }
     else {
         _autoValueBounds = true;
     }
 
-    if (dictionary.hasKeyAndValue<float>(KeyUpperValueBound)) {
-        _upperValueBound = dictionary.value<float>(KeyUpperValueBound);
+    if (dictionary.hasKeyAndValue<double>(KeyUpperValueBound)) {
+        _upperValueBound = static_cast<float>(
+            dictionary.value<double>(KeyUpperValueBound)
+        );
     }
     else {
         _autoValueBounds = true;
